Much better! The Mobile Vision API is inside.
The backend before reforming the UI is there.
There are functions to check if the device has camera/flash/autofocus.

The scanning and the options will be available only if the device has them (and prevent the ANR).
Finally caught the authors.split() bug. Very rarely there will be no information about it and unlike the rest, there is logic applied to it inside the AddBook and BookDetails.

The ISBN to test- 9780345340429 Lord of the Rings 4v

Well, almost done :)

To make it final a UI change will be needed.
Planed changes to the UI and structure is to make to add:
- Activities to the Fragments
- Fix the saveInstance states
- Make Book Cover imageview inside add book fragment fit with Glide (probably custom widht/height)
- Remove the dual-pane layout (it causes serious issues for tablet mode)
- Make changes to the tablet mode similar to the Materialized Sunshine app